+++ src/p_map.c
@@ -1752,9 +1752,30 @@ static boolean PTR_AimTraverse (intercept_t *in)
   return false;   // don't go any farther
 }
 
-// [Nugget] Explosive hitscan cheat /-----------------------------------------
+// [Nugget] /=================================================================
 
-boolean boomshot = false;
+// Extra Gibbing -------------------------------------------------------------
+
+static boolean is_bfg_tracer = false;
+
+boolean P_IsBFGTracer(void)
+{
+  return is_bfg_tracer;
+}
+
+void P_SetIsBFGTracer(const boolean value)
+{
+  is_bfg_tracer = value;
+}
+
+// Explosive hitscan cheat ---------------------------------------------------
+
+static boolean is_boomshot = false;
+
+void P_SetIsBoomShot(const boolean value)
+{
+  is_boomshot = value;
+}
 
 static void P_SpawnExplosion(fixed_t x, fixed_t y, fixed_t z)
 {
@@ -1765,7 +1786,7 @@ static void P_SpawnExplosion(fixed_t x, fixed_t y, fixed_t z)
   P_ExplodeMissile(mo);
 }
 
-// [Nugget] -----------------------------------------------------------------/
+// [Nugget] =================================================================/
 
 //
 // PTR_ShootTraverse
@@ -1854,8 +1875,11 @@ static boolean PTR_ShootTraverse(intercept_t *in)
       distance_travelled = P_AproxDistance(x - trace.x, y - trace.y); // [Nugget] Hitscan trails
 
       // [Nugget] Explosive hitscan cheat
-      if (boomshot)
+      if (is_boomshot)
+      {
+        is_boomshot = false;
         P_SpawnExplosion(x, y, z);
+      }
       else
     
       // Spawn bullet puffs.
@@ -1901,8 +1925,11 @@ static boolean PTR_ShootTraverse(intercept_t *in)
   distance_travelled = P_AproxDistance(x - trace.x, y - trace.y); // [Nugget] Hitscan trails
 
   // [Nugget] Explosive hitscan cheat
-  if (boomshot)
+  if (is_boomshot)
+  {
+    is_boomshot = false;
     P_SpawnExplosion(x, y, z);
+  }
   else
   // Spawn bullet puffs or blod spots,
   // depending on target type.
@@ -1997,7 +2024,7 @@ void P_LineAttack(mobj_t *t1, angle_t angle, fixed_t distance,
                         aimslope, attackrange, distance_travelled);
   }
 
-  boomshot = false; // [Nugget] Explosive hitscan cheat
+  is_boomshot = false; // [Nugget] Explosive hitscan cheat
 }
 
 //